delete option A benefit and embedded benefit cards
