+++ components/benefit_cards.js
@@ -2,60 +2,120 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { Grid, Typography, Button } from "@material-ui/core";
 import { withStyles } from "@material-ui/core/styles";
-import AddIcon from "@material-ui/icons/Add";
-import RemoveIcon from "@material-ui/icons/Remove";
-import EmbeddedBenefitCard from "./embedded_benefit_card";
+import ErrorOutlineIcon from "@material-ui/icons/ErrorOutline";
+import KeyboardReturnIcon from "@material-ui/icons/KeyboardReturn";
+import { KeyboardBackspace } from "@material-ui/icons";
+import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 import ExpansionPanel from "@material-ui/core/ExpansionPanel";
 import ExpansionPanelSummary from "@material-ui/core/ExpansionPanelSummary";
 import ExpansionPanelDetails from "@material-ui/core/ExpansionPanelDetails";
 import Highlighter from "react-highlight-words";
-import FavouriteButton from "./favourite_button";
+import FavouriteButton from "./favourite_button_b";
+import Paper from "@material-ui/core/Paper";
 import { logEvent } from "../utils/analytics";
 import { connect } from "react-redux";
 import NeedTag from "./need_tag";
+import EmbeddedBenefitCard from "./embedded_benefit_card";
 
 const styles = () => ({
   button: {
-    marginTop: "30px"
+    backgroundColor: "#3e57e2",
+    color: "white",
+    textAlign: "right"
+  },
+  cardBottom: {
+    backgroundColor: "#f1f7fc",
+    paddingLeft: "9px",
+    borderRadius: "0px",
+    borderTop: "1px solid #f5f5f5",
+    position: "relative"
+  },
+  cardBottomTitle: {
+    paddingLeft: "15px",
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center"
+  },
+  cardBottomFamilyTitle: {
+    marginLeft: "9px",
+    marginBottom: "25px",
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center"
+  },
+  CardBottomOpen: {
+    marginBottom: "0px",
+    marginTop: "0px",
+    "& $cardBottom": { backgroundColor: "#f5f5f5" }
+  },
+  cardTop: {
+    backgroundColor: "#f1f7fc",
+    borderRadius: "0px",
+    borderBottom: "1px solid #8b8b8b",
+    padding: "15px 15px 15px 10px",
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center"
+  },
+  cardBody: {
+    padding: "25px",
+    paddingTop: "15px"
   },
   cardDescriptionText: {
-    fontSize: "20px",
-    fontWeight: 400,
-    padding: "15px 0px"
+    fontSize: "18px",
+    padding: "10px 0px",
+    paddingBottom: "15px"
   },
   collapse: {
-    paddingTop: "25px"
+    paddingTop: "25px",
+    paddingLeft: "15px",
+    backgroundColor: "#f5f5f5"
   },
   root: {
     width: "100%"
   },
-  ExpansionPanelClosed: {
-    borderLeft: "5px solid"
-  },
-  ExpansionPanelOpen: {
-    borderLeft: "5px solid #808080"
-  },
-  ExpansionPanelSummary: {
-    "&[aria-expanded*=true]": {
-      backgroundColor: "#f8f8f8"
-    },
-    userSelect: "inherit"
-  },
-  ChildBenefitDesc: {
-    paddingBottom: "30px"
-  },
+  ExpansionPanelClosed: {},
   children: {
     width: "100%"
   },
-  ExampleDesc: {
-    paddingBottom: "10px"
-  },
-  examples: {
-    width: "100%",
-    marginLeft: "20px"
-  },
   benefitName: {
-    fontWeight: 500
+    fontWeight: 600,
+    padding: "10px 0"
+  },
+  returnIcon: {
+    "-moz-transform": "scaleX(-1)",
+    "-o-transform": "scaleX(-1)",
+    "-webkit-transform": "scaleX(-1)",
+    transform: "scaleX(-1)",
+    float: "left",
+    filter: "FlipH",
+    "-ms-filter": "FlipH",
+    paddingLeft: "10px"
+  },
+  rightArrowIcon: {
+    "-moz-transform": "scaleX(-1)",
+    "-o-transform": "scaleX(-1)",
+    "-webkit-transform": "scaleX(-1)",
+    transform: "scaleX(-1)",
+    float: "left",
+    filter: "FlipH",
+    "-ms-filter": "FlipH",
+    paddingRight: "10px"
+  },
+  parentIcon: {
+    flexGrow: 1,
+    marginRight: 15,
+    fontSize: 40,
+    transform: "scale(.9)",
+    color: "#434343"
+  },
+  headerDesc: {
+    width: "auto",
+    flexGrow: 3,
+    color: "#434343"
+  },
+  headerUrl: {
+    color: "#006CC9"
   }
 });
 
@@ -75,17 +135,85 @@ export class BenefitCard extends Component {
   };
 
   componentDidMount() {
-    this.props.onRef(this);
     this.forceUpdate();
   }
 
-  componentWillUnmount() {
-    this.props.onRef(undefined);
-  }
+  benefitTitle = benefit => {
+    return this.props.t("current-language-code") === "en"
+      ? benefit.vacNameEn
+      : benefit.vacNameFr;
+  };
+
+  benefitUrl = benefit => {
+    return this.props.t("current-language-code") === "en"
+      ? benefit.benefitPageEn
+      : benefit.benefitPageFr;
+  };
+
+  childBenefitNames = (benefit, childBenefits, open) => {
+    const length = childBenefits.length;
+    if (open) {
+      return this.props.t("benefits_b.eligible_open_veteran", {
+        x: this.benefitTitle(benefit)
+      });
+    } else {
+      if (length === 1) {
+        return (
+          this.benefitTitle(benefit) +
+          " " +
+          this.props.t("benefits_b.eligible_for_single", {
+            x: this.benefitTitle(childBenefits[0])
+          })
+        );
+      } else {
+        return (
+          this.benefitTitle(benefit) +
+          " " +
+          this.props.t("benefits_b.eligible_for_multi", {
+            x: length
+          })
+        );
+      }
+    }
+  };
+
+  get_benefit_a_elements = parentBenefits => {
+    let a_elements = parentBenefits.map((b, i) => (
+      <a
+        key={"a" + i}
+        className={this.props.classes.headerUrl}
+        href={this.benefitUrl(b)}
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        {this.benefitTitle(b)}
+      </a>
+    ));
+
+    let a_elements_with_ors = [];
+    a_elements.forEach((value, index) => {
+      if (a_elements.length - 1 !== index) {
+        a_elements_with_ors = a_elements_with_ors.concat([
+          value,
+          <span key={"b" + index}> {" " + this.props.t("index.or")} </span>
+        ]);
+      } else {
+        a_elements_with_ors = a_elements_with_ors.concat([
+          value,
+          <span key={"c" + index}> </span>
+        ]);
+      }
+    });
+    return a_elements_with_ors;
+  };
 
   render() {
     const benefit = this.props.benefit;
     const { t, classes } = this.props;
+
+    const parentBenefits = this.props.allBenefits.filter(
+      b => b.childBenefits && b.childBenefits.includes(benefit.id)
+    );
 
     const childBenefits = benefit.childBenefits
       ? this.props.allBenefits.filter(
@@ -100,12 +228,6 @@ export class BenefitCard extends Component {
       ab => this.props.familyBenefitIds.indexOf(ab.id) > -1
     );
 
-    const examples =
-      typeof benefit.examples !== "undefined" &&
-      typeof this.props.examples !== "undefined"
-        ? this.props.examples.filter(ex => benefit.examples.indexOf(ex.id) > -1)
-        : [];
-
     const needsMet = benefit.needs
       ? this.props.needs.filter(
           need =>
@@ -117,112 +239,129 @@ export class BenefitCard extends Component {
     return (
       <Grid item xs={12}>
         <div className={classes.root}>
-          <ExpansionPanel
-            className={
-              this.state.open
-                ? classes.ExpansionPanelOpen
-                : classes.ExpansionPanelClosed
-            }
-            expanded={this.state.open}
-          >
-            <ExpansionPanelSummary
-              className={classes.ExpansionPanelSummary}
-              expandIcon={this.state.open ? <RemoveIcon /> : <AddIcon />}
-              onClick={() => this.toggleOpenState()}
+          {parentBenefits.length > 0 &&
+          benefit.availableIndependently == "Requires Gateway Benefit" ? (
+            <Paper className={classes.cardTop}>
+              <ErrorOutlineIcon className={classes.parentIcon} />
+              <span className={classes.headerDesc}>
+                <span>{t("benefits_b.card_header_1") + " "}</span>
+                {this.get_benefit_a_elements(parentBenefits)}{" "}
+                <span>
+                  {this.props.t("benefits_b.card_header_2") +
+                    " " +
+                    this.benefitTitle(this.props.benefit) +
+                    "."}
+                </span>
+              </span>
+            </Paper>
+          ) : (
+            ""
+          )}
+
+          <Paper className={classes.cardBody}>
+            <div component="p" className={classes.benefitName}>
+              <Highlighter
+                searchWords={this.props.searchString.split(",")}
+                autoEscape={true}
+                textToHighlight={
+                  this.props.t("current-language-code") === "en"
+                    ? benefit.vacNameEn
+                    : benefit.vacNameFr
+                }
+              />
+            </div>
+
+            <Typography
+              className={"cardDescription " + classes.cardDescriptionText}
             >
-              <div>
-                <div component="p" className={classes.benefitName}>
-                  {this.props.showFavourite ? (
-                    <FavouriteButton
-                      benefit={benefit}
-                      toggleOpenState={this.toggleOpenState}
-                      store={this.props.store}
-                    />
-                  ) : (
-                    ""
-                  )}
-                  <Highlighter
-                    searchWords={this.props.searchString.split(",")}
-                    autoEscape={true}
-                    textToHighlight={
-                      this.props.t("current-language-code") === "en"
-                        ? benefit.vacNameEn
-                        : benefit.vacNameFr
-                    }
-                  />
-                  <div style={{ display: "inline-flex", marginLeft: "15px" }}>
-                    {needsMet.map(need => (
-                      <NeedTag key={benefit.id + need.id} t={t} need={need} />
-                    ))}
-                  </div>
+              <Highlighter
+                searchWords={this.props.searchString.split(",")}
+                autoEscape={true}
+                textToHighlight={
+                  this.props.t("current-language-code") === "en"
+                    ? benefit.oneLineDescriptionEn
+                    : benefit.oneLineDescriptionFr
+                }
+              />
+            </Typography>
+            <div>
+              {needsMet.map(need => (
+                <NeedTag
+                  key={benefit.id + need.id}
+                  t={this.props.t}
+                  need={need}
+                />
+              ))}
+            </div>
+
+            {this.props.showFavourite ? (
+              <FavouriteButton
+                benefit={benefit}
+                toggleOpenState={() => {}}
+                store={this.props.store}
+                t={this.props.t}
+              />
+            ) : (
+              ""
+            )}
+            <Button
+              className={classes.button}
+              target="_blank"
+              variant="raised"
+              style={{ textTransform: "none", float: "right" }}
+              onClick={() =>
+                this.logExit(
+                  this.props.t("current-language-code") === "en"
+                    ? benefit.benefitPageEn
+                    : benefit.benefitPageFr
+                )
+              }
+              href={
+                this.props.t("current-language-code") === "en"
+                  ? benefit.benefitPageEn
+                  : benefit.benefitPageFr
+              }
+            >
+              {this.props.t("Find out more")}
+              <KeyboardBackspace className={classes.rightArrowIcon} />
+            </Button>
+          </Paper>
+
+          {childBenefits.length > 0 ? (
+            <ExpansionPanel
+              expanded={this.state.open}
+              className={
+                this.state.open
+                  ? classes.CardBottomOpen
+                  : classes.ExpansionPanelClosed
+              }
+            >
+              <ExpansionPanelSummary
+                className={classes.cardBottom}
+                expandIcon={<ExpandMoreIcon />}
+                onClick={() => this.toggleOpenState()}
+              >
+                <div className={classes.cardBottomTitle}>
+                  <KeyboardReturnIcon className={classes.returnIcon} />
+                  <span className={classes.headerDesc}>
+                    <span>
+                      {this.childBenefitNames(
+                        benefit,
+                        childBenefits,
+                        this.state.open
+                      )}
+                    </span>
+                  </span>
                 </div>
+              </ExpansionPanelSummary>
 
-                <Typography
-                  className={"cardDescription " + classes.cardDescriptionText}
-                >
-                  <Highlighter
-                    searchWords={this.props.searchString.split(",")}
-                    autoEscape={true}
-                    textToHighlight={
-                      this.props.t("current-language-code") === "en"
-                        ? benefit.oneLineDescriptionEn
-                        : benefit.oneLineDescriptionFr
-                    }
-                  />
-                </Typography>
-              </div>
-            </ExpansionPanelSummary>
-
-            <ExpansionPanelDetails timeout="auto" className={classes.collapse}>
-              <Grid container spacing={24}>
-                <Grid item xs={12}>
-                  {examples.length > 0 ? (
-                    <Typography className={classes.ExampleDesc}>
-                      {t("examples") + ":"}
-                    </Typography>
-                  ) : (
-                    ""
-                  )}
-                  <Typography className={classes.examples}>
-                    {examples.map(ex => {
-                      return (
-                        <li key={ex.id}>
-                          {this.props.t("current-language-code") === "en"
-                            ? ex.nameEn
-                            : ex.nameFr}{" "}
-                        </li>
-                      );
-                    })}
-                  </Typography>
-
-                  <Button
-                    className={classes.button}
-                    target="_blank"
-                    variant="raised"
-                    style={{ textTransform: "none" }}
-                    onClick={() =>
-                      this.logExit(
-                        this.props.t("current-language-code") === "en"
-                          ? benefit.benefitPageEn
-                          : benefit.benefitPageFr
-                      )
-                    }
-                    href={
-                      this.props.t("current-language-code") === "en"
-                        ? benefit.benefitPageEn
-                        : benefit.benefitPageFr
-                    }
-                  >
-                    {this.props.t("Find out more")}
-                  </Button>
-                </Grid>
-
+              <ExpansionPanelDetails
+                timeout="auto"
+                className={classes.collapse}
+              >
                 <Grid item xs={12}>
                   {veteranBenefits.length > 0 ? (
                     <div className={classes.children}>
-                      <Typography className={classes.ChildBenefitDesc}>
-                        {t("Veteran child benefits")}:
-                      </Typography>
                       <div>
                         {veteranBenefits.map((cb, i) => (
                           <EmbeddedBenefitCard
@@ -230,7 +369,6 @@ export class BenefitCard extends Component {
                             benefit={cb}
                             t={this.props.t}
                             key={cb.id}
-                            onRef={ref => this.children.push(ref)}
                             showFavourite={this.props.showFavourite}
                             store={this.props.store}
                           />
@@ -244,11 +382,14 @@ export class BenefitCard extends Component {
                   )}
 
                   {familyBenefits.length > 0 ? (
-                    <div className={classes.children}>
-                      <Typography className={classes.ChildBenefitDesc}>
-                        {t("Family child benefits")}:
-                      </Typography>
-                      <div>
+                    <div>
+                      <div className={classes.cardBottomFamilyTitle}>
+                        <KeyboardReturnIcon className={classes.returnIcon} />
+                        <span className={classes.headerDesc}>
+                          {t("benefits_b.eligible_open_family")}
+                        </span>
+                      </div>
+                      <div className={classes.children}>
                         {familyBenefits.map((cb, i) => (
                           <EmbeddedBenefitCard
                             id={"cb" + i}
@@ -256,7 +397,6 @@ export class BenefitCard extends Component {
                             benefit={cb}
                             t={this.props.t}
                             key={cb.id}
-                            onRef={ref => this.children.push(ref)}
                             showFavourite={this.props.showFavourite}
                             store={this.props.store}
                           />
@@ -267,9 +407,11 @@ export class BenefitCard extends Component {
                     ""
                   )}
                 </Grid>
-              </Grid>
-            </ExpansionPanelDetails>
-          </ExpansionPanel>
+              </ExpansionPanelDetails>
+            </ExpansionPanel>
+          ) : (
+            ""
+          )}
         </div>
       </Grid>
     );
@@ -278,23 +420,22 @@ export class BenefitCard extends Component {
 
 const mapStateToProps = reduxState => {
   return {
-    examples: reduxState.examples,
     needs: reduxState.needs,
-    selectedNeeds: reduxState.selectedNeeds
+    selectedNeeds: reduxState.selectedNeeds,
+    benefits: reduxState.benefits
   };
 };
 
 BenefitCard.propTypes = {
+  benefits: PropTypes.array.isRequired,
   allBenefits: PropTypes.array.isRequired,
   veteranBenefitIds: PropTypes.array.isRequired,
   familyBenefitIds: PropTypes.array.isRequired,
   benefit: PropTypes.object.isRequired,
   classes: PropTypes.object.isRequired,
-  examples: PropTypes.array.isRequired,
   needs: PropTypes.array.isRequired,
   selectedNeeds: PropTypes.object.isRequired,
   t: PropTypes.func.isRequired,
-  onRef: PropTypes.func.isRequired,
   showFavourite: PropTypes.bool.isRequired,
   searchString: PropTypes.string.isRequired,
   store: PropTypes.object
