+++ components/embedded_benefit_card.js
@@ -3,46 +3,39 @@ import PropTypes from "prop-types";
 import { Typography, Button, Grid } from "@material-ui/core";
 import classnames from "classnames";
 import { withStyles } from "@material-ui/core/styles";
-import ExpansionPanel from "@material-ui/core/ExpansionPanel";
-import ExpansionPanelSummary from "@material-ui/core/ExpansionPanelSummary";
-import ExpansionPanelDetails from "@material-ui/core/ExpansionPanelDetails";
-import AddIcon from "@material-ui/icons/Add";
-import RemoveIcon from "@material-ui/icons/Remove";
-import FavouriteButton from "./favourite_button";
-import NeedTag from "./need_tag";
+import { KeyboardBackspace } from "@material-ui/icons";
 import { logEvent } from "../utils/analytics";
-import { connect } from "react-redux";
+import Paper from "@material-ui/core/Paper";
 
 const styles = theme => ({
   root: {
-    width: "100%"
+    margin: "20px",
+    marginTop: "0px",
+    padding: "20px",
+    paddingTop: "5px"
   },
   heading: {
+    paddingLeft: "0px",
+    paddingRight: "0px",
+    textTransform: "none",
     fontSize: theme.typography.pxToRem(15),
-    fontWeight: 500
-  },
-  ExpansionPanelClosed: {
-    borderLeft: "5px solid"
-  },
-  ExpansionPanelOpen: {
-    borderLeft: "5px solid #808080"
-  },
-  ExpansionPanelSummary: {
-    "&[aria-expanded*=true]": {
-      backgroundColor: "#f8f8f8"
-    },
-    userSelect: "inherit"
-  },
-  bullet: {
-    paddingBottom: "1em"
-  },
-  description: {
-    paddingTop: "1em"
+    fontWeight: 500,
+    color: "#3c51e6"
   },
   cardDescriptionText: {
-    fontSize: "20px",
-    fontWeight: 400,
-    padding: "15px 0px"
+    fontSize: "15px",
+    lineHeight: "1.6",
+    paddingTop: "0px"
+  },
+  rightArrowIcon: {
+    "-moz-transform": "scaleX(-1)",
+    "-o-transform": "scaleX(-1)",
+    "-webkit-transform": "scaleX(-1)",
+    transform: "scaleX(-1)",
+    float: "left",
+    filter: "FlipH",
+    "-ms-filter": "FlipH",
+    paddingRight: "5px"
   }
 });
 
@@ -61,124 +54,51 @@ export class EmbeddedBenefitCard extends Component {
     });
   };
 
-  componentDidMount() {
-    this.props.onRef(this);
-    const needsMet = this.getNeedsMet();
-    if (needsMet.length > 0) {
-      this.setState({ open: true });
-    }
-  }
-
-  componentWillUnmount() {
-    this.props.onRef(undefined);
-  }
-
-  getNeedsMet() {
-    return this.props.benefit.needs
-      ? this.props.needs.filter(
-          need =>
-            this.props.benefit.needs.indexOf(need.id) > -1 &&
-            this.props.selectedNeeds[need.id]
-        )
-      : [];
-  }
-
   render() {
     const { t, classes, benefit } = this.props;
     const language = t("current-language-code");
-    const needsMet = this.getNeedsMet();
 
     return (
-      <ExpansionPanel
-        className={
-          this.state.open
-            ? classes.ExpansionPanelOpen
-            : classes.ExpansionPanelClosed
-        }
-        expanded={this.state.open}
-      >
-        <ExpansionPanelSummary
-          expandIcon={this.state.open ? <RemoveIcon /> : <AddIcon />}
-          onClick={() => this.toggleOpenState()}
-          className={classes.ExpansionPanelSummary}
+      <Paper className={classes.root}>
+        <Button
+          target="_blank"
+          className={classes.heading}
+          onClick={() =>
+            this.logExit(
+              language === "en" ? benefit.benefitPageEn : benefit.benefitPageFr
+            )
+          }
+          href={
+            language === "en" ? benefit.benefitPageEn : benefit.benefitPageFr
+          }
         >
-          <div className={classes.heading}>
-            {this.props.showFavourite ? (
-              <FavouriteButton
-                benefit={benefit}
-                toggleOpenState={this.toggleOpenState}
-                store={this.props.store}
-              />
-            ) : (
-              ""
-            )}
-            {language === "en" ? benefit.vacNameEn : benefit.vacNameFr}
-            <div style={{ display: "inline-flex", paddingLeft: "15px" }}>
-              {needsMet.map(need => (
-                <NeedTag key={benefit.id + need.id} t={t} need={need} />
-              ))}
-            </div>
-          </div>
-        </ExpansionPanelSummary>
-        <ExpansionPanelDetails>
-          <Grid container spacing={24}>
-            <Grid item xs={12}>
-              <Typography
-                variant="title"
-                className={classnames(classes.cardDescriptionText)}
-              >
-                {language === "en"
-                  ? benefit.oneLineDescriptionEn
-                  : benefit.oneLineDescriptionFr}
-              </Typography>
-            </Grid>
-            <Grid item xs={12}>
-              <Button
-                size="small"
-                target="_blank"
-                variant="raised"
-                style={{ textTransform: "none" }}
-                onClick={() =>
-                  this.logExit(
-                    language === "en"
-                      ? benefit.benefitPageEn
-                      : benefit.benefitPageFr
-                  )
-                }
-                href={
-                  language === "en"
-                    ? benefit.benefitPageEn
-                    : benefit.benefitPageFr
-                }
-              >
-                {t("Find out more")}
-              </Button>
-            </Grid>
+          {language === "en" ? benefit.vacNameEn : benefit.vacNameFr}
+          <KeyboardBackspace className={classes.rightArrowIcon} />
+        </Button>
+
+        <Grid container spacing={24}>
+          <Grid item xs={12}>
+            <Typography
+              variant="title"
+              className={classnames(classes.cardDescriptionText)}
+            >
+              {language === "en"
+                ? benefit.oneLineDescriptionEn
+                : benefit.oneLineDescriptionFr}
+            </Typography>
           </Grid>
-        </ExpansionPanelDetails>
-      </ExpansionPanel>
+        </Grid>
+      </Paper>
     );
   }
 }
 
-const mapStateToProps = reduxState => {
-  return {
-    needs: reduxState.needs,
-    selectedNeeds: reduxState.selectedNeeds
-  };
-};
-
 EmbeddedBenefitCard.propTypes = {
   benefit: PropTypes.object.isRequired,
-  needs: PropTypes.array.isRequired,
-  selectedNeeds: PropTypes.object.isRequired,
   classes: PropTypes.object.isRequired,
   t: PropTypes.func.isRequired,
-  onRef: PropTypes.func.isRequired,
   showFavourite: PropTypes.bool.isRequired,
   store: PropTypes.object
 };
 
-export default connect(mapStateToProps)(
-  withStyles(styles)(EmbeddedBenefitCard)
-);
+export default withStyles(styles)(EmbeddedBenefitCard);
